Move some files on desktop and mobile folders for the navbar
