+++ src/theme/components/material-kit/Navbar/desktop/DropdownMenu.tsx
@@ -19,10 +19,10 @@ import React, { Dispatch, Fragment, SetStateAction, useState } from 'react';
 import { Box, Divider, Grid, Grow, Popper, Theme } from '@mui/material';
 import { ArrowDropUp, KeyboardArrowDown } from '@mui/icons-material';
 
-import { HeaderMenu, HeaderRoute } from '../../../../content/HeaderRoutes';
-import { MKTypography } from '../MKTypography';
-import { MKBox, MKBoxProps } from '../MKBox';
-import { getLinkAttributes } from '../../Link';
+import { HeaderMenu, HeaderRoute } from '../../../../../content/HeaderRoutes';
+import { MKTypography } from '../../MKTypography';
+import { MKBox, MKBoxProps } from '../../MKBox';
+import { getLinkAttributes } from '../../../Link';
 
 export const DropdownMenu = ({
   routes,
