+++ src/theme/components/material-kit/Navbar/desktop/NestedDropdownMenu.tsx
@@ -18,10 +18,10 @@ import React, { Dispatch, SetStateAction } from 'react';
 import { Grow, Popper, Theme } from '@mui/material';
 import { KeyboardArrowDown } from '@mui/icons-material';
 
-import { HeaderRoute } from '../../../../content/HeaderRoutes';
-import { MKTypography } from '../MKTypography';
-import { MKBox, MKBoxProps } from '../MKBox';
-import { getLinkAttributes } from '../../Link';
+import { HeaderRoute } from '../../../../../content/HeaderRoutes';
+import { MKTypography } from '../../MKTypography';
+import { MKBox, MKBoxProps } from '../../MKBox';
+import { getLinkAttributes } from '../../../Link';
 
 export const NestedDropdownMenu = ({
   routes,
